@@ -47,24 +47,23 @@
     "e2e": "ng e2e"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "~0.13.0",
-    "@angular/cli": "~7.3.5",
-    "@angular/common": "^7.2.8",
-    "@angular/compiler": "^7.2.8",
-    "@angular/compiler-cli": "^7.2.8",
-    "@angular/core": "^7.2.8",
-    "@angular/forms": "^7.2.8",
-    "@angular/http": "^7.2.8",
-    "@angular/language-service": "^7.2.8",
-    "@angular/platform-browser": "^7.2.8",
-    "@angular/platform-browser-dynamic": "^7.2.8",
-    "@angular/router": "^7.2.8",
+    "@angular-devkit/build-angular": "~0.802.1",
+    "@angular-devkit/build-ng-packagr": "~0.802.1",
+    "@angular/cli": "~8.2.1",
+    "@angular/common": "^8.2.1",
+    "@angular/compiler": "^8.2.1",
+    "@angular/compiler-cli": "^8.2.1",
+    "@angular/core": "^8.2.1",
+    "@angular/forms": "^8.2.1",
+    "@angular/language-service": "^8.2.1",
+    "@angular/platform-browser": "^8.2.1",
+    "@angular/platform-browser-dynamic": "^8.2.1",
+    "@angular/router": "^8.2.1",
     "@earshinov/extract-scss-variables": "^0.1.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
+    "codelyzer": "^5.1.0",
     "core-js": "^2.5.4",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -73,19 +72,17 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.1.0",
+    "ng-packagr": "^5.3.0",
     "node-sass": "^4.12.0",
     "protractor": "~5.4.0",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.2",
     "scss-bundle": "^2.5.0",
     "ts-node": "~7.0.0",
-    "tsickle": ">=0.34.0",
+    "tsickle": "^0.36.0",
     "tslib": "^1.9.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.4",
-    "zone.js": "~0.8.29"
+    "typescript": "~3.5.3",
+    "zone.js": "~0.9.1"
   },
-  "peerDependencies": {
-    "@angular/common": ">= 5"
-  }
+  "peerDependencies": {}
 }
