@@ -41,7 +41,7 @@ import {WizardComponent} from '../components/wizard.component';
 })
 class WizardTestComponent {
 
-  @ViewChild(WizardComponent)
+  @ViewChild(WizardComponent, {static: false})
   public wizard: WizardComponent;
 
   public goToSecondStep = 1;
