@@ -28,7 +28,7 @@ import { checkWizardState } from '../util/test-utils';
 })
 class WizardTestComponent {
 
-  @ViewChild(WizardComponent)
+  @ViewChild(WizardComponent, {static: false})
   public wizard: WizardComponent;
 
   public eventLog: Array<string> = [];
