@@ -24,7 +24,7 @@ import {WizardComponent} from '../components/wizard.component';
   `
 })
 class WizardTestComponent {
-  @ViewChild(WizardComponent)
+  @ViewChild(WizardComponent, {static: false})
   public wizard: WizardComponent;
 }
 
